requirements.txt
Django>=1.5,<1.6
South
lxml>=2.2.8
progressbar>=2.2
configparser>=3.2.0
pytz
djangorestframework>=2.3.10,<3.0.0
IPy

# required for operation on linux
dbus
daemon
